// src/models/Appointment.ts
export default interface Appointment {
    fechaCreacionTurnoUTC: string,
    fechaUltimaModificacionUTC: string,
    fechaTurnoUTC: string,
    email: string,
    celular?: string,
    servicio: string,
    prestador: string,
    ubicacion: string,
    observaciones: string,
    estado: string,
    id: string, // auto-generated by us as an internal id.
    fichaPacienteId?: number, // patient record ID for reference
    turnoId?: number // generated by DASPU's backend when the appointment is Confirmed (from Pending to Confirmed)
}